refactor: move book_form_live
