--- apps/app_web/lib/app_web/live/book_live/form.html.heex
@@ -1,29 +1,0 @@
-<.page_header back_to={@back_to}>
-  <:title>
-    <%= if @live_action == :new,
-      do: gettext("New Book"),
-      else: gettext("Edit Book") %>
-  </:title>
-</.page_header>
-
-<main class="mx-4">
-  <.form :let={f} for={@changeset} id="book-form" phx-change="validate" phx-submit="save">
-    <.input type="text" label={gettext("Name")} field={f[:name]} />
-
-    <.label>
-      <%= gettext("How to balance by default?") %>
-      <% # TODO Set value from changeset %>
-      <select name="book[default_balance_params][means_code]">
-        <option value="divide_equally"><%= gettext("Divide equally") %></option>
-        <option value="weight_by_income"><%= gettext("Weight by income") %></option>
-      </select>
-      <.error :for={msg <- f[:default_balance_params].errors}><%= msg %></.error>
-    </.label>
-
-    <div>
-      <.button color={:cta} phx-disable-with={gettext("Saving...")}>
-        <%= gettext("Save") %>
-      </.button>
-    </div>
-  </.form>
-</main>
